feat: add support for real time streaming of TTS Audio
We've also cleaned up some cruft and removed accidentally committed mp3
file. This should not happen from now on as we've updated .gitignore
to prevent this and other audio files from being committed to git.
