+++ examples/tts-stream-url/main.go
@@ -21,7 +21,7 @@ func main() {
 
 	// Creates an API client with default options.
 	// * it reads PLAYHT_SECRET_KEY and PLAYHT_USER_ID env vars
-	// * uses playht.BaserURL and APIv2
+	// * uses playht.BaserURL and APIv2 to create API endpoint URL
 	client := playht.NewClient()
 
 	voices, err := client.GetVoices(context.Background())
@@ -35,27 +35,20 @@ func main() {
 
 	voice := voices[0].ID
 
-	req := &playht.CreateTTSJobReq{
+	req := &playht.CreateTTSStreamReq{
 		Text:         input,
 		Voice:        voice,
 		Quality:      playht.Low,
 		OutputFormat: playht.Mp3,
 		Speed:        1.0,
 		SampleRate:   24000,
-		VoiceEngine:  playht.PlayHTv2,
+		VoiceEngine:  playht.PlayHTv2Turbo,
 	}
 
-	job, err := client.CreateTTSJob(context.Background(), req)
+	streamURL, err := client.TTSStreamURL(context.Background(), req)
 	if err != nil {
-		log.Fatalf("failed to create a TTS job: %v", err)
+		log.Fatalf("failed to create stream URL: %v", err)
 	}
 
-	log.Printf("successfully created a new TTS job: %#v", job)
-
-	jobInfo, err := client.GetTTSJob(context.Background(), job.ID)
-	if err != nil {
-		log.Fatalf("failed getting %v job info: %v", job.ID, err)
-	}
-
-	log.Printf("successfully got job info: %#v", jobInfo)
+	log.Printf("successfully retrieved stream URL: %#v", streamURL)
 }
